backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
SQLAlchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.8.3
